// frontend/src/sections/patients/detail/patient.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type PatientVariables = {
  storeId: string;
  patientId: string;
};

export type PatientResult = {
  patient: {
  id: string;
  name: string;
  code: string;
  code2: string | null;
  firstName: string | null;
  lastName: string | null;
  gender: "FEMALE" | "MALE" | "TRANSGENDER" | "TRANSGENDER_MALE" | "TRANSGENDER_MALE_HORMONE" | "TRANSGENDER_MALE_SURGICAL" | "TRANSGENDER_FEMALE" | "TRANSGENDER_FEMALE_HORMONE" | "TRANSGENDER_FEMALE_SURGICAL" | "UNKNOWN" | "NON_BINARY" | null;
  dateOfBirth: string | null;
  age: number | null;
  phone: string | null;
  address1: string | null;
  address2: string | null;
  country: string | null;
  email: string | null;
  website: string | null;
  isDeceased: boolean;
  dateOfDeath: string | null;
  nextOfKinId: string | null;
  nextOfKinName: string | null;
  createdDatetime: string | null;
  customFields: unknown | null;
  nextOfKin: {
  id: string;
  name: string;
} | null;
} | null;
};

export const Patient = {
  query: "query patient($storeId: String!, $patientId: String!) {\n  patient(storeId: $storeId, patientId: $patientId) {\n    id\n    name\n    code\n    code2\n    firstName\n    lastName\n    gender\n    dateOfBirth\n    age\n    phone\n    address1\n    address2\n    country\n    email\n    website\n    isDeceased\n    dateOfDeath\n    nextOfKinId\n    nextOfKinName\n    createdDatetime\n    customFields\n    nextOfKin {\n      id\n      name\n    }\n  }\n}",
} as TypedDocument<PatientResult, PatientVariables>;